pyproject.toml
[tool.poetry]
name = "wmbrbot"
version = "0.1.0"
description = ""
authors = ["Kyle Anthony Williams <kyle.anthony.williams2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.0"
icalendar = "^5.0.5"
python-dotenv = "^1.0.0"
discord-py = {version = "^2.2.3", extras = ["voice"]}
selectolax = "^0.3.13"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
